examples/PubSub/SeatPositionController/requirements.txt
requests
dapr-ext-grpc-dev >= 1.4.0rc1.dev
dapr-dev >= 1.4.0rc1.dev